# openapi.yaml
openapi: 3.1.0
info:
  title: Gong Sales Intelligence API
  description: Query Gong call recordings and search transcripts semantically.
  version: 3.0.0
servers:
  - url: https://hubspot-gpt-api-w80a.onrender.com
paths:
  /gong/calls/search:
    post:
      operationId: searchGongCalls
      summary: Search Gong calls by date range
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                from_date:
                  type: string
                  description: Start date (YYYY-MM-DD). Defaults to 7 days ago.
                to_date:
                  type: string
                  description: End date (YYYY-MM-DD). Defaults to today.
                limit:
                  type: integer
                  description: Max results to return. Defaults to 20.
      responses:
        "200":
          description: List of calls
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  results:
                    type: array
                    items:
                      type: object
                      properties:
                        call_id:
                          type: string
                        title:
                          type: string
                        started:
                          type: string
                        duration_seconds:
                          type: number
                        url:
                          type: string
                        participants:
                          type: array
                          items:
                            type: string
                        direction:
                          type: string

  /gong/calls/{call_id}/transcript:
    get:
      operationId: getCallTranscript
      summary: Get full transcript of a Gong call
      parameters:
        - name: call_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Call transcript broken down by speaker segments
          content:
            application/json:
              schema:
                type: object
                properties:
                  call_id:
                    type: string
                  transcript:
                    type: array
                    items:
                      type: object
                      properties:
                        speaker_id:
                          type: string
                        topic:
                          type: string
                        text:
                          type: string
                        start_time:
                          type: number
                        end_time:
                          type: number
        "404":
          description: Transcript not found

  /gong/calls/{call_id}/stats:
    get:
      operationId: getCallStats
      summary: Get call statistics
      parameters:
        - name: call_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Call stats
          content:
            application/json:
              schema:
                type: object
                properties:
                  call_id:
                    type: string
                  title:
                    type: string
                  url:
                    type: string
                  duration_seconds:
                    type: number
                  started:
                    type: string
                  direction:
                    type: string
                  participants:
                    type: array
                    items:
                      type: string
        "404":
          description: Call not found

  /gong/contacts/{email}/calls:
    get:
      operationId: getContactCalls
      summary: Get all Gong calls for a specific contact email
      parameters:
        - name: email
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of calls for the contact
          content:
            application/json:
              schema:
                type: object
                properties:
                  email:
                    type: string
                  total_calls:
                    type: integer
                  calls:
                    type: array
                    items:
                      type: object
                      properties:
                        call_id:
                          type: string
                        title:
                          type: string
                        date:
                          type: string
                        duration_seconds:
                          type: number
                        url:
                          type: string

  /gong/search:
    post:
      operationId: searchTranscripts
      summary: Semantic search across all Gong transcripts. Use this to find calls mentioning specific topics, stories, phrases, or patterns.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - query
              properties:
                query:
                  type: string
                  description: Natural language search query (e.g. "Derek story", "pricing objection", "competitor mention")
                from_date:
                  type: string
                  description: Optional start date filter (YYYY-MM-DD)
                to_date:
                  type: string
                  description: Optional end date filter (YYYY-MM-DD)
                limit:
                  type: integer
                  description: Max results to return. Defaults to 20.
      responses:
        "200":
          description: Matching transcript chunks with call metadata
          content:
            application/json:
              schema:
                type: object
                properties:
                  query:
                    type: string
                  matching_chunks:
                    type: integer
                  unique_calls:
                    type: integer
                  results:
                    type: array
                    items:
                      type: object
                      properties:
                        call_id:
                          type: string
                        call_title:
                          type: string
                        call_date:
                          type: string
                        call_url:
                          type: string
                        speaker_id:
                          type: string
                        topic:
                          type: string
                        text:
                          type: string
                          description: The transcript snippet that matched the query
                        start_time:
                          type: number
                        end_time:
                          type: number
                        score:
                          type: number
                          description: Relevance score (0 to 1, higher is more relevant)